📖 Storybook:  TableRow 스토리 수정
